Further organize supabase logic
