+++ lib/pages/auth/login_select.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 
-import '../../supabase.dart';
+import '../../supabase/sso_auth.dart';
 import 'email_login.dart';
 
 class LoginSelectPage extends StatelessWidget {
